Move tests to right places and consolidate helpers
