+++ embrace-android-sdk/src/integrationTest/kotlin/io/embrace/android/embracesdk/testcases/session/UserSessionLifecycleTest.kt
@@ -9,8 +9,12 @@ import io.embrace.android.embracesdk.assertions.isFinalSessionPart
 import io.embrace.android.embracesdk.internal.arch.state.AppState
 import io.embrace.android.embracesdk.internal.config.remote.BackgroundActivityRemoteConfig
 import io.embrace.android.embracesdk.internal.config.remote.RemoteConfig
+import io.embrace.android.embracesdk.internal.config.remote.SessionRemoteConfig
 import io.embrace.android.embracesdk.internal.config.remote.UserSessionRemoteConfig
 import io.embrace.android.embracesdk.internal.otel.sdk.findAttributeValue
+import io.embrace.android.embracesdk.internal.payload.Envelope
+import io.embrace.android.embracesdk.internal.payload.SessionPartPayload
+import io.embrace.android.embracesdk.internal.worker.Worker
 import io.embrace.android.embracesdk.semconv.EmbSessionAttributes.EMB_SESSION_PART_ID
 import io.embrace.android.embracesdk.semconv.EmbSessionAttributes.EMB_USER_SESSION_ID
 import io.embrace.android.embracesdk.semconv.EmbSessionAttributes.EMB_USER_SESSION_NUMBER
@@ -19,6 +23,8 @@ import io.embrace.android.embracesdk.semconv.EmbSessionAttributes.EMB_USER_SESSI
 import io.embrace.android.embracesdk.semconv.EmbSessionAttributes.EmbUserSessionTerminationReasonValues.INACTIVITY
 import io.embrace.android.embracesdk.semconv.EmbSessionAttributes.EmbUserSessionTerminationReasonValues.MANUAL
 import io.embrace.android.embracesdk.testframework.SdkIntegrationTestRule
+import io.embrace.android.embracesdk.testframework.SdkIntegrationTestRule.Companion.DEFAULT_SDK_START_TIME_MS
+import io.embrace.android.embracesdk.testframework.actions.EmbraceSetupInterface
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotEquals
@@ -28,6 +34,7 @@ import org.junit.Assert.assertTrue
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import kotlin.time.Duration.Companion.seconds
 
 /**
  * Asserts that the user session obeys rules about its lifecycle
@@ -37,7 +44,13 @@ internal class UserSessionLifecycleTest {
 
     @Rule
     @JvmField
-    val testRule: SdkIntegrationTestRule = SdkIntegrationTestRule()
+    val testRule: SdkIntegrationTestRule = SdkIntegrationTestRule {
+        EmbraceSetupInterface(
+            workersToFake = listOf(Worker.Background.NonIoRegWorker),
+        ).apply {
+            getFakedWorkerExecutor(Worker.Background.NonIoRegWorker).blockingMode = false
+        }
+    }
 
     @Test
     fun `user session created only once the app enters the foreground`() {
@@ -195,11 +208,11 @@ internal class UserSessionLifecycleTest {
     @Test
     fun `load persisted user session in foreground state`() {
         val persistedId = "aabbccdd11223344aabbccdd11223344"
-        val startMs = SdkIntegrationTestRule.DEFAULT_SDK_START_TIME_MS - 1_000L
-        val lastActivityMs = SdkIntegrationTestRule.DEFAULT_SDK_START_TIME_MS - 100L
+        val startMs = DEFAULT_SDK_START_TIME_MS - 1_000L
+        val lastActivityMs = DEFAULT_SDK_START_TIME_MS - 100L
         testRule.runTest(
             setupAction = {
-                persistUserSession(sessionId = persistedId, startMs = startMs, lastActivityMs = lastActivityMs)
+                persistUserSession(userSessionId = persistedId, startMs = startMs, lastActivityMs = lastActivityMs)
             },
             testCaseAction = { recordSession() },
             assertAction = {
@@ -215,13 +228,13 @@ internal class UserSessionLifecycleTest {
     @Test
     fun `load persisted user session in inactivity timeout state`() {
         val persistedId = "aabbccdd11223344aabbccdd11223344"
-        val startMs = SdkIntegrationTestRule.DEFAULT_SDK_START_TIME_MS - 3_000_000L
+        val startMs = DEFAULT_SDK_START_TIME_MS - 3_000_000L
         // lastActivityMs more than inactivity timeout before clock.now(), so the session is inactive
         val defaultInactivityMs = 1800L * 1_000L
-        val lastActivityMs = SdkIntegrationTestRule.DEFAULT_SDK_START_TIME_MS - defaultInactivityMs - 1L
+        val lastActivityMs = DEFAULT_SDK_START_TIME_MS - defaultInactivityMs - 1L
         testRule.runTest(
             setupAction = {
-                persistUserSession(sessionId = persistedId, startMs = startMs, lastActivityMs = lastActivityMs)
+                persistUserSession(userSessionId = persistedId, startMs = startMs, lastActivityMs = lastActivityMs)
             },
             testCaseAction = { recordSession() },
             assertAction = {
@@ -239,12 +252,12 @@ internal class UserSessionLifecycleTest {
         val persistedId = "aabbccdd11223344aabbccdd11223344"
         // startMs more than max duration before clock.now(), so the session exceeds max duration
         val defaultMaxDurationMs = 43200L * 1_000L
-        val startMs = SdkIntegrationTestRule.DEFAULT_SDK_START_TIME_MS - defaultMaxDurationMs - 1L
+        val startMs = DEFAULT_SDK_START_TIME_MS - defaultMaxDurationMs - 1L
         // lastActivityMs is recent so inactivity check does not fire first
-        val lastActivityMs = SdkIntegrationTestRule.DEFAULT_SDK_START_TIME_MS - 1_000L
+        val lastActivityMs = DEFAULT_SDK_START_TIME_MS - 1_000L
         testRule.runTest(
             setupAction = {
-                persistUserSession(sessionId = persistedId, startMs = startMs, lastActivityMs = lastActivityMs)
+                persistUserSession(userSessionId = persistedId, startMs = startMs, lastActivityMs = lastActivityMs)
             },
             testCaseAction = { recordSession() },
             assertAction = {
@@ -290,5 +303,148 @@ internal class UserSessionLifecycleTest {
                 assertEquals("2", secondSession.attributes?.findAttributeValue(EMB_USER_SESSION_NUMBER))
             }
         )
+    }
+
+    @Test
+    fun `manual session end kill switch does not suppress timer-driven inactivity session transition`() {
+        val inactivityTimeoutSeconds = 30
+        testRule.runTest(
+            persistedRemoteConfig = RemoteConfig(
+                sessionConfig = SessionRemoteConfig(isEnabled = true),
+                userSession = UserSessionRemoteConfig(inactivityTimeoutSeconds = inactivityTimeoutSeconds),
+            ),
+            testCaseAction = {
+                recordSession()
+                clock.tick(inactivityTimeoutSeconds * 1_000L + 1L)
+                unblockTimerThread()
+                recordSession()
+            },
+            assertAction = {
+                val sessions = getSessionEnvelopes(2)
+                assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+            },
+        )
+    }
+
+    @Test
+    fun `inactivity timer at exact boundary expires the user session`() {
+        val inactivityTimeoutSeconds = 30
+        testRule.runTest(
+            persistedRemoteConfig = RemoteConfig(
+                userSession = UserSessionRemoteConfig(inactivityTimeoutSeconds = inactivityTimeoutSeconds),
+            ),
+            testCaseAction = {
+                recordSession()
+                clock.tick(inactivityTimeoutSeconds * 1_000L)
+                unblockTimerThread()
+                recordSession()
+            },
+            assertAction = {
+                val sessions = getSessionEnvelopes(2)
+                assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+            },
+        )
+    }
+
+    @Test
+    fun `max duration timer at exact boundary expires the user session`() {
+        val maxDurationSeconds = 3600
+        testRule.runTest(
+            persistedRemoteConfig = RemoteConfig(
+                userSession = UserSessionRemoteConfig(maxDurationSeconds = maxDurationSeconds),
+            ),
+            testCaseAction = {
+                recordSession()
+                clock.tick(maxDurationSeconds * 1_000L)
+                unblockTimerThread()
+                recordSession()
+            },
+            assertAction = {
+                val sessions = getSessionEnvelopes(2)
+                assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+            },
+        )
+    }
+
+    @Test
+    fun `manual end after max duration but before the job is processed terminates with MANUAL`() {
+        val maxDurationSeconds = 600
+        testRule.runTest(
+            persistedRemoteConfig = RemoteConfig(
+                userSession = UserSessionRemoteConfig(maxDurationSeconds = maxDurationSeconds),
+            ),
+            testCaseAction = {
+                recordSession {
+                    clock.tick(maxDurationSeconds * 1_000L)
+                    embrace.endUserSession()
+                }
+                unblockTimerThread()
+                recordSession()
+            },
+            assertAction = {
+                val sessions = getSessionEnvelopes(2)
+                assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+                assertEquals(MANUAL, sessions[0].getUserSessionTerminationReason())
+            },
+        )
+    }
+
+    @Test
+    fun `foregrounding before inactivity timeout continues the same user session`() {
+        runPartsBoundaryTest(
+            inactivityTimeoutSeconds = 30,
+            partsGapSeconds = 29
+        ) { sessions ->
+            assertEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+        }
+    }
+
+    @Test
+    fun `foregrounding at inactivity boundary creates a new user session`() {
+        runPartsBoundaryTest(
+            inactivityTimeoutSeconds = 60,
+            partsGapSeconds = 60
+        ) { sessions ->
+            assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+        }
+    }
+
+    @Test
+    fun `foregrounding after inactivity boundary creates a new user session`() {
+        runPartsBoundaryTest(
+            inactivityTimeoutSeconds = 30,
+            partsGapSeconds = 31
+        ) { sessions ->
+            assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+        }
+    }
+
+    private fun runPartsBoundaryTest(
+        inactivityTimeoutSeconds: Int,
+        partsGapSeconds: Int = 0,
+        assertions: (sessions: List<Envelope<SessionPartPayload>>) -> Unit,
+    ) {
+        testRule.runTest(
+            persistedRemoteConfig = RemoteConfig(
+                userSession = UserSessionRemoteConfig(inactivityTimeoutSeconds = inactivityTimeoutSeconds),
+            ),
+            testCaseAction = {
+                recordSession()
+                clock.tick(partsGapSeconds.seconds.inWholeMilliseconds)
+                recordSession()
+            },
+            assertAction = {
+                val sessions = getSessionEnvelopes(2)
+                assertions(sessions)
+                val reason = sessions[0].getUserSessionTerminationReason()
+                if (reason != null && partsGapSeconds >= inactivityTimeoutSeconds) {
+                    assertEquals(INACTIVITY, reason)
+                }
+            },
+        )
+    }
+
+    private fun unblockTimerThread() {
+        testRule.setup.getFakedWorkerExecutor(Worker.Background.NonIoRegWorker).runCurrentlyBlocked()
     }
 }
